requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
boto3==1.28.52
botocore==1.31.52
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.17.0
